--- .babelrc
@@ -1,11 +1,0 @@
-{
-  "presets": [
-    ["@babel/env"],
-    [
-      "@babel/typescript",
-      {
-        "allowDeclareFields": false
-      }
-    ]
-  ]
-}
