+++ packages/fabric/index.ts
@@ -1,0 +1,10 @@
+/// <reference path="./types/fabric.d.ts" />
+
+export { ClippingGroup } from './src/ClippingGroup';
+export {
+  EraserBrush,
+  type ErasingEndEvent,
+  type ErasingEndEventDetail,
+  eraseObject,
+  eraseCanvasDrawable,
+} from './src/EraserBrush';
