+++ packages/fabric/src/ClippingGroup.ts
