+++ packages/fabric/src/ErasingEffect.ts
