+++ packages/fabric/src/util.ts
