+++ packages/sandbox/README.md
@@ -7,3 +7,7 @@ A lean parcel app for testing and bug reporting.
 Provide a **detailed description** including steps to reproduce in the [`REPRODUCE.md`](./REPRODUCE.md) file.
 
 Navigate to [`src/index`](./src/index.ts) and start editing.
+
+## Dev
+
+see main [`README`](../../README.md).
