+++ rollup.config.mjs
@@ -1,16 +1,20 @@
 import { babel } from '@rollup/plugin-babel';
 import terser from '@rollup/plugin-terser';
 import ts from '@rollup/plugin-typescript';
+import path from 'node:path';
+import del from 'rollup-plugin-delete';
 
-// https://rollupjs.org/guide/en/#configuration-files
+const fabric = path.join('./packages', 'fabric');
+
 /**
+ * https://rollupjs.org/guide/en/#configuration-files
  * @type {import('rollup').RollupOptions}
  */
 export default [
   {
-    input: './src/fabric/index.ts',
+    input: path.join(fabric, 'index.ts'),
     output: {
-      dir: './dist',
+      dir: path.join(fabric, 'dist'),
       format: 'es',
       preserveModules: true,
       entryFileNames: '[name].js',
@@ -18,14 +22,18 @@ export default [
       plugins: [terser()],
     },
     plugins: [
+      del({
+        targets: [path.join(fabric, 'dist/*')],
+      }),
       ts({
         noForceEmit: true,
-        tsconfig: './tsconfig.json',
-        exclude: ['dist', '**/**.spec.ts', '**/**.test.ts'],
+        tsconfig: path.join(fabric, 'tsconfig.json'),
+        exclude: ['**/dist', '**/**.spec.ts', '**/**.test.ts'],
       }),
       babel({
         extensions: ['.ts', '.js'],
         babelHelpers: 'bundled',
+        presets: [['@babel/env'], ['@babel/typescript']],
       }),
     ],
     external: ['fabric'],
