--- src/fabric/index.ts
@@ -1,8 +1,0 @@
-export { ClippingGroup } from './ClippingGroup';
-export {
-  EraserBrush,
-  type ErasingEndEvent,
-  type ErasingEndEventDetail,
-  eraseObject,
-  eraseCanvasDrawable,
-} from './EraserBrush';
